[#77]; refactor: bag관련 물리로직 따로 분리
